Fix ioctl prototypes
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@7437 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/stored/vtape.c
@@ -998,7 +998,7 @@ void vtape::dump()
 
 #else  /* USE_VTAPE */
 
-int vtape_ioctl(int fd, unsigned long int request, ...)
+int vtape_ioctl(int fd, int request, ...)
 {
    return -1;
 }
